Version 2.5 (Check for updates and new settings and other fixes and improvements)
