+++ StatusCal/Constants.swift
@@ -1,0 +1,21 @@
+//
+//  Constants.swift
+//  StatusCal
+//
+//  Created by Alexey Boldakov on 22.04.2020.
+//  Copyright © 2020 Alex Boldakov. All rights reserved.
+//
+
+
+
+
+struct Constants {
+    static let launcherAppId = "ru.alexvr.StatusCalLauncher"
+    static let baseURL = "https://ru-webhost.ru"
+    
+    
+    
+    static let lastVersionURL = "https://raw.githubusercontent.com/AlexeyBoldakov/StatusCal/master/lastversion.json"
+    static let lastVersionDownloadURL = "https://github.com/AlexeyBoldakov/StatusCal/raw/master/builds/StatusCal.dmg.zip"
+    static let donateURL = "https://paypal.me/boldakov"
+}
