+++ StatusCal/Models/UpdateState.swift
@@ -1,0 +1,16 @@
+//
+//  UpdateState.swift
+//  StatusCal
+//
+//  Created by Alexey Boldakov on 23.04.2020.
+//  Copyright © 2020 Alex Boldakov. All rights reserved.
+//
+
+import Foundation
+
+
+enum UpdateState {
+    case startCheck
+    case newVersionAvailable
+    case currentIsLatest
+}
