+++ StatusCal/en.lproj/Localizable.strings
@@ -1,0 +1,13 @@
+/* 
+  Localizable.strings
+  StatusCal
+
+  Created by Alexey Boldakov on 24.04.2020.
+  Copyright © 2020 Alex Boldakov. All rights reserved.
+*/
+
+
+/* Update controller */
+"updates.checking.message" = "Checking for updates...";
+"updates.checking.available" = "New version is available now!";
+"updates.checking.newest" = "Сurrent version is the newest";
