+++ StatusCal/en.lproj/MainMenu.strings
@@ -1,68 +1,43 @@
-/* Class = "NSWindow"; title = "Settings"; ObjectID = "2fn-5w-bKg"; */
-"2fn-5w-bKg.title" = "Settings";
 
-/* Class = "NSWindow"; title = "About"; ObjectID = "jdE-c0-rAD"; */
-"jdE-c0-rAD.title" = "About";
-
-/* Class = "NSButtonCell"; title = "September 2015"; ObjectID = "5Js-yS-1ZZ"; */
-"5Js-yS-1ZZ.title" = "September 2015";
-
-/* Class = "NSButtonCell"; title = "Show date"; ObjectID = "BDg-ZQ-zFx"; */
-"BDg-ZQ-zFx.title" = "Show date";
-
-/* Class = "NSButtonCell"; title = "◀"; ObjectID = "JkT-H5-zCl"; */
-"JkT-H5-zCl.title" = "◀";
-
-/* Class = "NSButtonCell"; title = "Show the day of the week"; ObjectID = "SJf-Zh-g4M"; */
-"SJf-Zh-g4M.title" = "Show the day of the week";
-
-/* Class = "NSTextFieldCell"; title = "Time options"; ObjectID = "V7g-VS-7tR"; */
-"V7g-VS-7tR.title" = "Time options";
-
-/* Class = "NSButtonCell"; title = "Use a 24-hour clock"; ObjectID = "X0B-xC-XWg"; */
-"X0B-xC-XWg.title" = "Use a 24-hour clock";
-
-/* Class = "NSButtonCell"; title = "Show the time with seconds"; ObjectID = "YBS-XV-zIF"; */
-"YBS-XV-zIF.title" = "Show the time with seconds";
-
-/* Class = "NSTextFieldCell"; title = "Date options"; ObjectID = "fRr-jF-hAT"; */
-"fRr-jF-hAT.title" = "Date options";
-
-/* Class = "NSButtonCell"; title = "▶"; ObjectID = "gys-YA-Nr4"; */
-"gys-YA-Nr4.title" = "▶";
-
-/* Class = "NSButtonCell"; title = "Show AM/PM"; ObjectID = "wHM-Vz-n6I"; */
-"wHM-Vz-n6I.title" = "Show AM/PM";
-
-/* Class = "NSMenuItem"; title = "[Calendar Collection]"; ObjectID = "VJE-7c-auX"; */
-"VJE-7c-auX.title" = "[Calendar Collection]";
-
-/* Class = "NSMenuItem"; title = "Support this project"; ObjectID = "9E8-Hb-g0V"; */
-"9E8-Hb-g0V.title" = "Support this project";
-
-/* Class = "NSMenuItem"; title = "Settings"; ObjectID = "q73-Mh-ATb"; */
-"q73-Mh-ATb.title" = "Open calendar app";
-
-/* Class = "NSMenuItem"; title = "Settings"; ObjectID = "0GA-Pk-dni"; */
+/* Root menu */
 "0GA-Pk-dni.title" = "Settings";
-
-/* Class = "NSMenuItem"; title = "About"; ObjectID = "5Ab-Tp-fCl"; */
+"q73-Mh-ATb.title" = "Open calendar app";
 "5Ab-Tp-fCl.title" = "About";
-
-/* Class = "NSMenuItem"; title = "Quit"; ObjectID = "nwf-du-ktX"; */
+"9E8-Hb-g0V.title" = "Support this project";
+"Imf-7V-9Mt.title" = "Check for updates";
 "nwf-du-ktX.title" = "Quit";
 
-/* Class = "NSTextField"; title = "Version"; ObjectID = "b6y-2e-ZIm"; */
-"b6y-2e-ZIm.title" = "Version";
+/* Calendar */
+"VJE-7c-auX.title" = "[Calendar Collection]";
+"5Js-yS-1ZZ.title" = "September 2015";
+"JkT-H5-zCl.title" = "◀";
+"gys-YA-Nr4.title" = "▶";
 
-/* Class = "NSTextField"; title = "Copyright © 2020 Alex Boldakov. All rights reserved."; ObjectID = "3yw-kd-nA2"; */
-"3yw-kd-nA2.text" = "Copyright © 2020 Alex Boldakov. All rights reserved.";
 
-/* Class = "NSTabViewItem"; title = "Time options"; ObjectID = "ljO-fF-h9p"; */
+/* Settings */
+"2fn-5w-bKg.title" = "Settings";
+"UML-DR-xC2.label" = "General";
+"nZO-xo-XSf.title" = "Launch at login";
 "ljO-fF-h9p.label" = "Time options";
-
-/* Class = "NSTabViewItem"; title = "Calendar options"; ObjectID = "beM-E2-q1p"; */
+"V7g-VS-7tR.title" = "Time options";
+"fRr-jF-hAT.title" = "Date options";
+"BDg-ZQ-zFx.title" = "Show date";
+"SJf-Zh-g4M.title" = "Show the day of the week";
+"X0B-xC-XWg.title" = "Use a 24-hour clock";
+"YBS-XV-zIF.title" = "Show the time with seconds";
+"wHM-Vz-n6I.title" = "Show AM/PM";
 "beM-E2-q1p.label" = "Calendar options";
-
-/* Class = "NSButtonCell"; title = "Show calendar events"; ObjectID = "WXx-sH-f08"; */
 "WXx-sH-f08.title" = "Show calendar events";
+
+
+/* About */
+"jdE-c0-rAD.title" = "About";
+"b6y-2e-ZIm.title" = "Version";
+"04C-Hw-fMT.title" = "Copyright © 2020 Alex Boldakov. All rights reserved.";
+"saC-ne-fH1.title" = "Project page";
+
+
+/* Updates */
+"Zcv-LP-zGA.title" = "Download latest";
+
+
